Create StringTokens
INPUT:A single string of s
OUTPUT:On the first line, print an integer,n , denoting the number of tokens in string s (they do not need to be unique). Next, print each of the n tokens on a new line in the same order as they appear in input string s.